Cvs 112872 enabling examples (#20)
* Remove unused graphs

* Early enabling

* iris wip

* Iris fixed.

* Readme

* Remove face_mesh target

* Add force flag

+++ setup_ovms.py
@@ -24,16 +24,16 @@ import traceback
 __version__ = '1.0'
 
 class SetupOpenvinoModelServer():
-  def __init__(self):
+  def __init__(self, force):
       self.build_lib = "mediapipe/models/ovms"
+      self.force = force
 
   def run_command(self, command):
     print(command)
     if subprocess.call(command.split()) != 0:
       sys.exit(-1)
 
-  def _copy_to_build_lib_dir(self, build_lib, file):
-    """Copy a file from bazel-bin to the build lib dir."""
+  def get_dst(self, build_lib, file):
     dst = os.path.join(build_lib + '/', file.replace("/","/1/"))
     dst_dir = os.path.dirname(file)
     # Workaround to copy every model in separate directory
@@ -51,6 +51,13 @@ class SetupOpenvinoModelServer():
        build_file = os.path.join('mediapipe/modules/', file)
 
     dst_dir = os.path.dirname(dst)
+
+    return dst, dst_dir, build_file
+  
+  def _copy_to_build_lib_dir(self, build_lib, file):
+    """Copy a file from bazel-bin to the build lib dir."""
+    
+    dst, dst_dir, build_file = self.get_dst(build_lib, file)
     if not os.path.exists(dst_dir):
       os.makedirs(dst_dir)
     print("Copy to: " + dst)
@@ -97,11 +104,15 @@ class SetupOpenvinoModelServer():
     shutil.copyfile(file_to_copy, dst)
 
   def convert_pose(self):
+    dst = "mediapipe/models/ovms/pose_detection/1/pose_detection.tflite"
+    if os.path.exists(dst) and not self.force:
+      print("file exists, not converting: " + dst + " use --force argument to overwrite.\n")
+      return
     print("Converting pose detection model")
-    self.run_command("cp -r  mediapipe/models/ovms/pose_detection/1/pose_detection.tflite .")
+    self.run_command("cp -r  " + dst +" .")
     self.run_command("tflite2tensorflow --model_path pose_detection.tflite --flatc_path flatbuffers/build/flatc --schema_path schema.fbs --output_pb")
     self.run_command("tflite2tensorflow --model_path pose_detection.tflite --flatc_path flatbuffers/build/flatc --schema_path schema.fbs --output_no_quant_float32_tflite   --output_dynamic_range_quant_tflite   --output_weight_quant_tflite   --output_float16_quant_tflite   --output_integer_quant_tflite")
-    self.run_command("cp -rf saved_model/model_float32.tflite mediapipe/models/ovms/pose_detection/1/pose_detection.tflite")
+    self.run_command("cp -rf saved_model/model_float32.tflite " + dst)
     self.run_command("rm -rf pose_detection.tflite")
     self.run_command("rm -rf saved_model")
 
@@ -116,7 +127,7 @@ class SetupOpenvinoModelServer():
         'pose_landmark/pose_landmark_by_roi_cpu.pbtxt',
     ]
     for elem in external_files:
-      sys.stderr.write('coping file: %s\n' % elem)
+      print('coping file: %s\n' % elem)
       self._copy_pbxt_file(elem)
 
   def get_models(self):
@@ -144,7 +155,11 @@ class SetupOpenvinoModelServer():
         'models/ssdlite_object_detection.tflite',
     ]
     for elem in external_files:
-      sys.stderr.write('downloading file: %s\n' % elem)
+      dst, dst_dir, build_lib = self.get_dst(self.build_lib, elem)
+      if os.path.exists(dst) and not self.force:
+         sys.stderr.write("file exists, not downloading: " + dst  + " use --force argument to overwrite.\n")
+         continue
+      print('downloading file: %s\n' % elem)
       self._download_external_file(elem)
 
 def printUsage():
@@ -170,8 +185,9 @@ def get_args(argv):
     get_graphs_flag = False
     get_models_flag = False
     convert_pose = False
+    force = False
     try:
-        opts, vals = getopt.getopt(argv, "", ["convert_pose","get_graphs","get_models","help"])
+        opts, vals = getopt.getopt(argv, "", ["force","convert_pose","get_graphs","get_models","help"])
     except getopt.GetoptError:
         print("ERROR: unrecognize option/missing argument/value for known option. Use --help to see list of options")
         sys.exit(2)
@@ -185,17 +201,19 @@ def get_args(argv):
           get_models_flag = True
         elif opt in ("--convert_pose"):
           convert_pose = True
+        elif opt in ("--force"):
+          force = True
 
-    return get_graphs_flag, get_models_flag, convert_pose
+    return get_graphs_flag, get_models_flag, convert_pose, force
 
 if __name__ == "__main__":
-  get_graphs_flag, get_models_flag, convert_pose = get_args(sys.argv[1:])
+  get_graphs_flag, get_models_flag, convert_pose, force = get_args(sys.argv[1:])
   if get_models_flag:
-    SetupOpenvinoModelServer().get_models()
+    SetupOpenvinoModelServer(force).get_models()
 
   # Needed to call only on starting ovm holistic demo from ovms repository using ovms server standalone instance
   if get_graphs_flag:
-    SetupOpenvinoModelServer().get_graphs()
+    SetupOpenvinoModelServer(force).get_graphs()
 
   if convert_pose:
-    SetupOpenvinoModelServer().convert_pose()
+    SetupOpenvinoModelServer(force).convert_pose()
